Connection over tcp and udp working
